Picodst fixes (#585)
Leszek reported (on the simulations list) an issue with converting
simulation MuDsts to picoDsts. The code issues a couple of errors, and
then crashes. I believe this is due to a faulty assumption in
initialization. The code relies on the filename begining with "st_" in
order to determine the basename of the file, when the full path is
given. This fails for Leszek's case (and for any pure simualtion). When
the assumption fails, the initialization routine returns before creating
the TTree. This results in the crash during Make, when the null pointer
to the TTree is deferenced.

This PR removes the faulty assumption, and makes early returns from
Init() a "fatal" return rather than just an error.